+++ index.html
@@ -463,7 +463,7 @@
       </div>
     </template>
 
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
 
   </body>
 </html>
