+++ js/getRandomNumber.js
@@ -1,0 +1,9 @@
+const getRandomNumber = function (min, max) {
+  if (max > min && min >= 0 && max > 0) {
+    return Math.floor(min + Math.random() * (max + 1 - min));
+  }
+
+  throw new RangeError('error here');
+};
+
+export {getRandomNumber};
