+++ js/isUnique.js
@@ -1,0 +1,3 @@
+const isUnique = (array, value) => !array.includes(value);
+
+export {isUnique};
